Restructure Hermione favors and small fixes
reorganised the structure of how you would read through the favors attempting ro remove unnecessary jumps
attempting to structure things in a way which you would expect them to progress in game i.e. the topmost
if block in most sisutations would be expected to be the firt "scene" or "event" or if its just text then
 the more lewd the text the further down the if blocks it goes indacateing progression

also attempting to create a uniform way to address if whoring or equivalent value in in a given range i.e.

if lower_bound <= value <= upper_bound:

this is to try to create an easy to read range that is difficult to mistake for anything else